+++ Modules/Options/Options.lua
@@ -521,6 +521,7 @@ function addon:OnEnable()
 				{ "autolooting_text", "description" },
 				{ "speedy_autoloot", "toggle", width = "double" },
 				{ "speedy_autoloot_respect_filters", "toggle", width = "double", requires = "speedy_autoloot" },
+				{ "speedy_autoloot_skip_fishing", "toggle", width = "double", requires = "speedy_autoloot" },
 				{ "autoloot_currency", when_group, "autoloots", "currency" },
 				{ "autoloot_quest", when_group, "autoloots", "quest" },
 				{ "autoloot_tradegoods", when_group, "autoloots", "tradegoods" },